+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - Created 1.2.2 build
 
 ### Changed
 
 ### Fixed
+- Fixed a crash on selection of prims compatible with particle systems
+- Fixed logic to enable particle system creation on prims
 
 ### Removed