+++ source/extensions/lightspeed.trex.stage_manager.plugin.widget.usd/docs/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.2.2]
+### Fixed
+- Fixed logic to enable particle system creation on prims
+
+## [2.2.1]
+### Fixed
+- Fixed a bug on selection of prims compatible with particle systems
+
 ## [2.2.0]
 ### Added
 - Added an icon for the particle system action